Additional changes related with Sidebar
